wd_requirements.txt
wheel
torch
onnxruntime-gpu
pandas
numpy
pillow
huggingface_hub
